# docs/swagger.yaml
definitions:
  global.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
      traceId:
        type: string
    type: object
...
host: 127.0.0.1:8080
info:
  contact:
    email: 25076778@qq.com
    name: API Support
    url: http://www.swagger.io/support
  description: Gin-base API 文档
  termsOfService: http://swagger.io/terms/
  title: Gin-base Swagger API
  version: "2.0"
paths:
...
  /api/v1/menu:
    get:
      description: 菜单列表
      parameters:
      - description: 认证Token
        in: header
        name: token
        required: true
        type: string
      - description: 更新参数
        in: body
        name: data
        required: true
        schema:
          type: object
      responses:
        "200":
          description: 成功返回" Example({"code":0,"msg":"Success","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "500":
          description: 系统错误" Example({"code":500,"msg":"系统错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
      summary: 列表
      tags:
      - 菜单管理
    post:
      description: 菜单创建
      parameters:
      - description: 认证Token
        in: header
        name: token
        required: true
        type: string
      - description: 创建参数
        in: body
        name: data
        required: true
        schema:
          type: object
      - description: 更新参数
        in: body
        name: data
        required: true
        schema:
          type: object
      responses:
        "200":
          description: 成功返回" Example({"code":0,"msg":"Success","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "400":
          description: 参数错误" Example({"code":400,"msg":"参数错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "500":
          description: 系统错误" Example({"code":500,"msg":"系统错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
      summary: 创建
      tags:
      - 菜单管理
  /api/v1/menu/{id}:
    delete:
      description: 菜单删除
      parameters:
      - description: 认证Token
        in: header
        name: token
        required: true
        type: string
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新参数
        in: body
        name: data
        required: true
        schema:
          type: object
      responses:
        "200":
          description: 成功返回" Example({"code":0,"msg":"Success","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "400":
          description: 参数错误" Example({"code":400,"msg":"参数错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "500":
          description: 系统错误" Example({"code":500,"msg":"系统错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
      summary: 删除
      tags:
      - 菜单管理
    get:
      description: 菜单详情
      parameters:
      - description: 认证Token
        in: header
        name: token
        required: true
        type: string
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新参数
        in: body
        name: data
        required: true
        schema:
          type: object
      responses:
        "200":
          description: 成功返回" Example({"code":0,"msg":"Success","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "400":
          description: 参数错误" Example({"code":400,"msg":"参数错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "500":
          description: 系统错误" Example({"code":500,"msg":"系统错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
      summary: 详情
      tags:
      - 菜单管理
    put:
      description: 菜单更新
      parameters:
      - description: 认证Token
        in: header
        name: token
        required: true
        type: string
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新参数
        in: body
        name: data
        required: true
        schema:
          type: object
      - description: 更新参数
        in: body
        name: data
        required: true
        schema:
          type: object
      responses:
        "200":
          description: 成功返回" Example({"code":0,"msg":"Success","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "400":
          description: 参数错误" Example({"code":400,"msg":"参数错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "500":
          description: 系统错误" Example({"code":500,"msg":"系统错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
      summary: 更新
      tags:
      - 菜单管理
  /api/v1/menu/{id}/action:
    get:
      description: 菜单功能列表
      parameters:
      - description: 认证Token
        in: header
        name: token
        required: true
        type: string
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新参数
        in: body
        name: data
        required: true
        schema:
          type: object
      responses:
        "200":
          description: 成功返回" Example({"code":0,"msg":"Success","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "400":
          description: 参数错误" Example({"code":400,"msg":"参数错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "500":
          description: 系统错误" Example({"code":500,"msg":"系统错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
      summary: 功能列表
      tags:
      - 菜单管理
    post:
      description: 菜单功能创建
      parameters:
      - description: 认证Token
        in: header
        name: token
        required: true
        type: string
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 创建参数
        in: body
        name: data
        required: true
        schema:
          type: object
      - description: 更新参数
        in: body
        name: data
        required: true
        schema:
          type: object
      responses:
        "200":
          description: 成功返回" Example({"code":0,"msg":"Success","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "400":
          description: 参数错误" Example({"code":400,"msg":"参数错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "500":
          description: 系统错误" Example({"code":500,"msg":"系统错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
      summary: 功能创建
      tags:
      - 菜单管理
  /api/v1/menu/{id}/action/{actionId}:
    delete:
      description: 菜单功能删除
      parameters:
      - description: 认证Token
        in: header
        name: token
        required: true
        type: string
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 功能ID
        in: path
        name: actionId
        required: true
        type: integer
      - description: 更新参数
        in: body
        name: data
        required: true
        schema:
          type: object
      responses:
        "200":
          description: 成功返回" Example({"code":0,"msg":"Success","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "400":
          description: 参数错误" Example({"code":400,"msg":"参数错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "500":
          description: 系统错误" Example({"code":500,"msg":"系统错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
      summary: 功能删除
      tags:
      - 菜单管理
    get:
      description: 菜单功能详情
      parameters:
      - description: 认证Token
        in: header
        name: token
        required: true
        type: string
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 功能ID
        in: path
        name: actionId
        required: true
        type: integer
      - description: 更新参数
        in: body
        name: data
        required: true
        schema:
          type: object
      responses:
        "200":
          description: 成功返回" Example({"code":0,"msg":"Success","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "400":
          description: 参数错误" Example({"code":400,"msg":"参数错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "500":
          description: 系统错误" Example({"code":500,"msg":"系统错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
      summary: 功能详情
      tags:
      - 菜单管理
    put:
      description: 菜单功能创建
      parameters:
      - description: 认证Token
        in: header
        name: token
        required: true
        type: string
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 功能ID
        in: path
        name: actionId
        required: true
        type: integer
      - description: 更新参数
        in: body
        name: data
        required: true
        schema:
          type: object
      - description: 更新参数
        in: body
        name: data
        required: true
        schema:
          type: object
      responses:
        "200":
          description: 成功返回" Example({"code":0,"msg":"Success","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "400":
          description: 参数错误" Example({"code":400,"msg":"参数错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "500":
          description: 系统错误" Example({"code":500,"msg":"系统错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
      summary: 功能创建
      tags:
      - 菜单管理
  /api/v1/menu/role-menu:
    get:
      description: 角色菜单
      parameters:
      - description: 认证Token
        in: header
        name: token
        required: true
        type: string
      - description: 角色ids
        in: query
        name: roleIds
        required: true
        type: string
      - description: 更新参数
        in: body
        name: data
        required: true
        schema:
          type: object
      responses:
        "200":
          description: 成功返回" Example({"code":0,"msg":"Success","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "500":
          description: 系统错误" Example({"code":500,"msg":"系统错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
      summary: 角色菜单
      tags:
      - 菜单管理
...
swagger: "2.0"